requirements.txt
discord.py
flask-cors
flask